refactor: useQuery select option를 api단에서 처리
