+++ packages/mobile/src/page/Notice/NoticeItem/style.module.scss
