+++ packages/mobile/src/page/Notice/NoticeList/style.module.scss
